Factor out CountDownLatch and Gate to separate header files.
