+++ searchcore/src/vespa/searchcore/proton/attribute/exclusive_attribute_read_accessor.cpp
@@ -1,6 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "exclusive_attribute_read_accessor.h"
+#include <vespa/vespalib/util/gate.h>
 #include <vespa/searchlib/attribute/attributevector.h>
 #include <vespa/searchlib/common/isequencedtaskexecutor.h>
