+++ vespalib/src/vespa/vespalib/util/sync.h
@@ -527,108 +527,5 @@ public:
     }
 };
 
-
-/**
- * A countdown latch helps one or more threads wait for the completion
- * of a number of operations performed by other threads. Specifically,
- * any thread invoking the await method will block until the countDown
- * method has been invoked an appropriate number of times. The
- * countdown latch is created with a count. Each invocation of
- * countDown will reduce the current count. When the count reaches 0,
- * the threads blocked in await will be unblocked. When the count is
- * 0, additional invocations of await will not block and additional
- * invocations of countDown will have no effect.
- **/
-class CountDownLatch
-{
-private:
-    Monitor  _monitor;
-    uint32_t _count;
-
-    CountDownLatch(const CountDownLatch &rhs) = delete;
-    CountDownLatch &operator=(const CountDownLatch &rhs) = delete;
-
-public:
-    /**
-     * Create a countdown latch with the given initial count.
-     *
-     * @param cnt initial count
-     **/
-    CountDownLatch(uint32_t cnt) : _monitor(), _count(cnt) {}
-
-    /**
-     * Count down this latch. When the count reaches 0, all threads
-     * blocked in the await method will be unblocked.
-     **/
-    void countDown() {
-        MonitorGuard guard(_monitor);
-        if (_count == 0) {
-            return;
-        }
-        --_count;
-        if (_count == 0) {
-            guard.broadcast();
-        }
-    }
-
-    /**
-     * Wait for this latch to count down to 0. This method will block
-     * until the countDown method has been invoked enough times to
-     * reduce the count to 0.
-     **/
-    void await() {
-        MonitorGuard guard(_monitor);
-        while (_count != 0) {
-            guard.wait();
-        }
-    }
-
-    /**
-     * Wait for this latch to count down to 0. This method will block
-     * until the countDown method has been invoked enough times to
-     * reduce the count to 0 or the given amount of time has elapsed.
-     *
-     * @param maxwait the maximum number of milliseconds to wait
-     * @return true if the counter reached 0, false if we timed out
-     **/
-    bool await(int maxwait) {
-        MonitorGuard guard(_monitor);
-        TimedWaiter waiter(guard, maxwait);
-        while (_count != 0 && waiter.hasTime()) {
-            waiter.wait();
-        }
-        return (_count == 0);
-    }
-
-    /**
-     * Obtain the current count for this latch. This method is mostly
-     * useful for debugging and testing.
-     *
-     * @return current count
-     **/
-    uint32_t getCount() const {
-        return _count;
-    }
-
-    /**
-     * Empty. Needs to be virtual to reduce compiler warnings.
-     **/
-    virtual ~CountDownLatch() = default;
-};
-
-
-/**
- * A gate is a countdown latch with an initial count of 1, indicating
- * that we are only waiting for a single operation to complete.
- **/
-class Gate : public CountDownLatch
-{
-public:
-    /**
-     * Sets the initial count to 1.
-     **/
-    Gate() : CountDownLatch(1) {}
-};
-
 } // namespace vespalib
