+++ vespalib/src/vespa/vespalib/util/thread.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "sync.h"
+#include "gate.h"
 #include "runnable.h"
 #include "active.h"
 #include <vespa/fastos/thread.h>
